feat: add events for initialize, swap, and modify (#115)
* feat: add events for initialize, swap, and modify

This commit adds a few events for key flows in PoolManager - initialize,
swap, and modify. Note gas tradeoffs are unfortunately quite heavy, and
I omitted a few useful arguments for `Swap` specifically as they would
require several extra sloads from the pool contract without major refactor,
like liquidity and new sqrtprice

Fix code style issues with Prettier

* Fix code style issues with Prettier

* fix: review comments

* feat: return sqrtPriceX96 from Pool.swap

This commit adds sqrtPriceX96 as a return from swap, and in the Swap log

* fix: only hash poolId once

This commit removes duplicate hashes of pool key to get the value that
PoolManager emits in ModifyPosition and Swap

* fix: review comments

- log pool id in initialize
- sqrtPriceX96 -> sqrtPriceX96After
- poolKey -> poolId
- remove sqrtPriceX96 from swapState since it is in the return

* fix: poolDefinition -> poolKey

* feat: add liquidity and tick to swap event

This commit adds the liquidity and tick fields to the swap event, for
event param parity with v3-core. To do this, we return the entire
SwapState struct from the Pool.swap() library call

* Fix code style issues with Prettier

Co-authored-by: Lint Action <lint-action@samuelmeuli.com>