Simple storage and memory caching optimizations in `Pool` (#567)
* Cache `tick` and `sqrtPriceX96` in `Pool.modifyLiquidity`

This commit slightly modifies the `Pool` library to reduce gas usage. An intermediary variable `tick` and `sqrtPriceX96` are created to reduce the number of calls to `self.slot0`. This change results in smaller footprint in the snapshots, indicating reduction in gas consumption.

* Cache `zeroForOne` in stack in `Pool.swap`

Parameters in Pool.sol are now more concisely referenced with `zeroForOne` variable. This affects all the occurrences of `params.zeroForOne`, where it is replaced with `zeroForOne`. The snapshot files are also updated to reflect this change. This simplification helps to reduce bytecode size and increase overall efficiency of the code.

* Cache `self.liquidity` and remove clone to memory in `Position.update`

This commit focuses on optimizing the `Position.update` function from within the `Position.sol` file. It introduces a cached variable `liquidity`, thereby eliminating the need to clone `self` to memory. This change results in a more efficient use of memory and optimizes code execution. The snapshots have been updated to reflect the adjusted bytecode sizes for the various operations.

* Cache `params.liquidityDelta` to stack in `Pool.modifyLiquidity`

This commit primarily caches the variable `params.liquidityDelta` to stack in the `Pool.modifyLiquidity` method inside Pool.sol to improve efficiency. This tiny optimization reduces bytecode size and makes the underlying calculations faster and more efficient. This change is reflected in the updated .snap files for poolManager bytecode size and various test cases involving liquidity management.

* Cache `tickLower` and `tickUpper` to stack in `Pool.modifyLiquidity`

The changes in the code have optimized memory use in the `Pool.modifyLiquidity` function, by caching `tickLower` and `tickUpper` to the stack rather than repeatedly accessing them from memory. This minor adjustment not only optimizes memory usage, but also potentially improves the function performance by reducing the number of memory operations.