Refactor and optimize `Pool.swap` (#603)
* Optimize tick transition in `Pool.swap`

Pool.sol has been updated with a block of assembly code for the assignment of `state.tick`. This is equivalent but slightly more efficient version of the previous code. Additionally, the snapshot values have been adjusted as a consequence of this change.

* Move `exactInput` and `step` declaration

* Update `while` condition in `Pool.swap`

The `while` condition within the `swap` function of the `Pool` contract has been changed for clarity and ease of understanding. The condition now explicitly checks for equality compared to the previous implementation. All corresponding snapshot files have been updated to reflect this change.

* Update tick boundaries conditions in `Pool.swap`

* Switch if-else branches in `Pool.swap`

This commit swaps the conditional branches in the `Pool.swap` function in `Pool.sol`. The switch leads to changes in calculated snapshots, indicating reduced gas.

* Refactor variable assignment in `Pool`

The refactoring of variable assignment in `Pool.sol` aims to increase code readability and maintainability. The change includes replacing separate lines of variable assignments with destructuring assignment, where multiple variables are assigned values in a single statement, making the code more concise and efficient.

* Refactor fee growth assignment in `Pool.swap`

This update simplifies the fee growth assignments in the `swap` method of the `Pool.sol` file. Previously, the fee growth assignments were performed inside a verbose conditional statement. The refactoring condenses this into a single line of code using a ternary operator, improving the code readability and maintainability.

* Optimize Pool.sol for reduced bytecode size

This commit includes an optimization in the Pool.sol code where a conditional statement was modified to reduce bytecode size. The snapshot of the bytecode size shows a slight decrease due to these changes, thus ensuring a more efficient contract deployment.