Optimize overflow check in `mulDivRoundingUp` (#667)
* Optimize overflow check in `mulDivRoundingUp`

This commit optimizes the overflow check within the `mulDivRoundingUp` function of FullMath library. This is achieved by combining two steps: the increment of the `result` and the overflow check, into a single step using the `++` operator. This simplifies the code, improves its readability and brings down the gas usage slightly, as reflected in updated snapshot files.

* Update assertions in `test_fuzz_mulDivRoundingUp`